pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reguq"
version = "0.1.0"
description = "Colab-first regression uncertainty quantification toolkit"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Danesh Selwal" }
]
dependencies = [
  "numpy==1.26.4",
  "pandas==2.2.3",
  "scikit-learn==1.6.1",
  "scipy==1.14.1",
  "PyYAML==6.0.2",
  "matplotlib==3.10.1",
  "optuna==4.6.0",
  "openpyxl==3.1.5",
  "XlsxWriter==3.2.9",
  "mapie==0.6.0",
  "puncc==0.8.0",
  "lightgbm==4.6.0",
  "xgboost==3.1.2",
  "catboost==1.2.8",
  "ngboost==0.5.8",
  "pgbm==2.2.0",
  "properscoring==0.1"
]

[project.optional-dependencies]
interpretability = [
  "shap==0.44.0",
  "lime==0.2.0.1",
  "interpret==0.7.4"
]
advanced_cp = []
card = []

[project.urls]
Homepage = "https://github.com/DaneshSelwal"
Repository = "https://github.com/DaneshSelwal/Regression_Uncertainty_Quantification_Analysis"
Issues = "https://github.com/DaneshSelwal/Regression_Uncertainty_Quantification_Analysis/issues"

[project.scripts]
reguq-run = "reguq.runner:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
